Validation & Extracting the Table Availability Logic Into it's Own Fucntion Release #13
