+++ pages/api/restaurant/[slug]/availability.ts
@@ -1,6 +1,6 @@
-import { NextApiRequest, NextApiResponse } from "next";
-import { times } from "../../../../data";
 import { PrismaClient } from "@prisma/client";
+import { NextApiRequest, NextApiResponse } from "next";
+import { findAvailabileTables } from "../../../../services/restaurant/findAvailableTables";
 
 const prisma = new PrismaClient();
 
@@ -22,41 +22,6 @@ export default async function handler(
       });
     }
 
-    const searchTimes = times.find((t) => {
-      return t.time === time;
-    })?.searchTimes;
-
-    if (!searchTimes) {
-      return res.status(400).json({
-        errorMessage: "Invalid provided",
-      });
-    }
-
-    const bookings = await prisma.booking.findMany({
-      where: {
-        booking_time: {
-          gte: new Date(`${day}T${searchTimes[0]}`),
-          lte: new Date(`${day}T${searchTimes[searchTimes.length - 1]}`),
-        },
-      },
-      select: {
-        number_of_people: true,
-        booking_time: true,
-        tables: true,
-      },
-    });
-
-    const bookingTablesObj: { [key: string]: { [key: number]: true } } = {};
-    bookings.forEach((booking) => {
-      bookingTablesObj[booking.booking_time.toISOString()] =
-        booking.tables.reduce((obj, table) => {
-          return {
-            ...obj,
-            [table.table_id]: true,
-          };
-        }, {});
-    });
-
     const restaurant = await prisma.restaurant.findUnique({
       where: {
         slug,
@@ -74,23 +39,18 @@ export default async function handler(
       });
     }
 
-    const tables = restaurant.tables;
-    const searchTimesWithTables = searchTimes.map((searchTime) => {
-      return {
-        date: new Date(`${day}T${searchTime}`),
-        time: searchTime,
-        tables,
-      };
+    const searchTimesWithTables = await findAvailabileTables({
+      day,
+      time,
+      res,
+      restaurant,
     });
 
-    searchTimesWithTables.forEach((t) => {
-      t.tables = t.tables.filter((table) => {
-        if (bookingTablesObj[t.date.toISOString()]) {
-          if (bookingTablesObj[t.date.toISOString()][table.id]) return false;
-        }
-        return true;
+    if (!searchTimesWithTables) {
+      return res.status(400).json({
+        errorMessage: "Invalid data provided",
       });
-    });
+    }
 
     const availabilities = searchTimesWithTables
       .map((t) => {
@@ -110,6 +70,7 @@ export default async function handler(
         const timeIsBeforeClosingHour =
           new Date(`${day}T${availability.time}`) <=
           new Date(`${day}T${restaurant.close_time}`);
+
         return timeIsAfterOpeningHour && timeIsBeforeClosingHour;
       });
 
@@ -117,5 +78,4 @@ export default async function handler(
   }
 }
 
-// http://localhost:3000/api/restaurant/vivaan-fine-indian-cuisine-ottawa/availability?day=2023-03-02&time=14:00:00.000Z&partySize=4
-// vivaan-fine-indian-cuisine-ottawa
+// http://localhost:3000/api/restaurant/vivaan-fine-indian-cuisine-ottawa/availability?day=2023-02-03&time=15:00:00.000Z&partySize=8
